feat(configManager): add getFromCache
It's not meant to be used a lot, just in place where it's much more
convenient and easy to rely on cache (no async). In theory, the cache
should always be up to date because we handle additions/deletions and
cache everything at the beginning. But we never know ;) better use get()
in most cases